Refactor database classes to their own package
