+++ android/src/main/java/com/thebluealliance/androidclient/adapters/EventCursorAdapter.java
@@ -11,7 +11,7 @@ import android.widget.TextView;
 
 import com.thebluealliance.androidclient.Constants;
 import com.thebluealliance.androidclient.R;
-import com.thebluealliance.androidclient.datafeed.Database;
+import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.helpers.EventHelper;
 
 import java.util.Date;
