+++ android/src/main/java/com/thebluealliance/androidclient/background/mytba/PopulateUserSubscriptions.java
@@ -14,7 +14,7 @@ import com.thebluealliance.androidclient.adapters.ListViewAdapter;
 import com.thebluealliance.androidclient.comparators.SubscriptionSortByModelComparator;
 import com.thebluealliance.androidclient.datafeed.APIResponse;
 import com.thebluealliance.androidclient.datafeed.DataManager;
-import com.thebluealliance.androidclient.datafeed.Database;
+import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.datafeed.RequestParams;
 import com.thebluealliance.androidclient.fragments.mytba.MySubscriptionsFragment;
 import com.thebluealliance.androidclient.helpers.AnalyticsHelper;
