+++ android/src/main/java/com/thebluealliance/androidclient/datafeed/TBAv2.java
@@ -10,6 +10,7 @@ import com.google.gson.JsonElement;
 import com.squareup.okhttp.Response;
 import com.thebluealliance.androidclient.Constants;
 import com.thebluealliance.androidclient.Utilities;
+import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.helpers.DistrictHelper;
 import com.thebluealliance.androidclient.models.District;
 import com.thebluealliance.androidclient.models.Event;
