+++ android/src/main/java/com/thebluealliance/androidclient/gcm/notifications/SummaryNotification.java
@@ -9,7 +9,7 @@ import android.support.v4.app.NotificationCompat;
 import com.google.gson.JsonParseException;
 import com.thebluealliance.androidclient.R;
 import com.thebluealliance.androidclient.activities.RecentNotificationsActivity;
-import com.thebluealliance.androidclient.datafeed.Database;
+import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.gcm.GCMMessageHandler;
 import com.thebluealliance.androidclient.listeners.NotificationDismissedListener;
 import com.thebluealliance.androidclient.models.StoredNotification;
